# docs/openapi/admin-api.yaml
openapi: "3.0.3"
info:
  title: "PJB Admin API"
  version: "1.0.0-frontend-freeze"
  description: "Contrato exportado estaticamente para integração com frontend."
servers:
  -
    url: "http://localhost:8080"
paths:
...
  /api/admin/judicial/crypto-ops/inspect/{system}/{tribunalCodigo}:
    get:
      summary: "AdminJudicialConnectorCryptoOpsController::GET /api/admin/judicial/crypto-ops/inspect/{system}/{tribunalCodigo}"
      tags:
        - "com.tcc.pjb.backend.controller.admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/admin/judicial/crypto-ops/probe/{system}/{tribunalCodigo}:
    get:
      summary: "AdminJudicialConnectorCryptoOpsController::GET /api/admin/judicial/crypto-ops/probe/{system}/{tribunalCodigo}"
      tags:
        - "com.tcc.pjb.backend.controller.admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
...
  /api/v1/admin/frontend-readiness/auth:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/auth"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/blockers:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/blockers"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/bootstrap:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/bootstrap"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/checklist:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/checklist"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/errors:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/errors"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/public-contract/dtos:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/public-contract/dtos"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/public-contract/envelopes:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/public-contract/envelopes"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/public-contract/error-catalog:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/public-contract/error-catalog"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/public-contract/freeze:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/public-contract/freeze"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/public-contract/http-freeze:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/public-contract/http-freeze"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/public-contract/routes:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/public-contract/routes"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/public-contract/summary:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/public-contract/summary"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/public-contract/validation:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/public-contract/validation"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/public-routes:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/public-routes"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/frontend-readiness/summary:
    get:
      summary: "AdminFrontendReadinessController::GET /api/v1/admin/frontend-readiness/summary"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
...
  /api/v1/admin/institucional/catalogo/cobertura-nacional:
    get:
      summary: "AdminInstitutionalCatalogGovernanceController::GET /api/v1/admin/institucional/catalogo/cobertura-nacional"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
...
  /api/v1/admin/institucional/catalogo/expandir-baseline-regional:
    post:
      summary: "AdminInstitutionalCatalogGovernanceController::POST /api/v1/admin/institucional/catalogo/expandir-baseline-regional"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInstitutionalRegionalBaselineExpansionRequest"
      security:
        -
          bearerAuth:
  /api/v1/admin/institucional/catalogo/governancas:
    get:
      summary: "AdminInstitutionalCatalogGovernanceController::GET /api/v1/admin/institucional/catalogo/governancas"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:

    post:
      summary: "AdminInstitutionalCatalogGovernanceController::POST /api/v1/admin/institucional/catalogo/governancas"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInstitutionalCatalogGovernanceUpsertRequest"
      security:
        -
          bearerAuth:
  /api/v1/admin/institucional/catalogo/regras-competencia:
    get:
      summary: "AdminInstitutionalCatalogGovernanceController::GET /api/v1/admin/institucional/catalogo/regras-competencia"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:

    post:
      summary: "AdminInstitutionalCatalogGovernanceController::POST /api/v1/admin/institucional/catalogo/regras-competencia"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInstitutionalCompetenceRuleUpsertRequest"
      security:
        -
          bearerAuth:
  /api/v1/admin/institucional/catalogo/sumario:
    get:
      summary: "AdminInstitutionalCatalogGovernanceController::GET /api/v1/admin/institucional/catalogo/sumario"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
    basicAuth:
      type: "http"
      scheme: "basic"